my-streamlit-app/requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
langchain_community
google-generativeai
sentence-transformers
pytesseract
faiss-cpu
beautifulsoup4